requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
cryptography==44.0.2
distro==1.9.0
dnspython==2.8.0
dotenv==0.9.9
firebase_admin==7.0.0
Flask==3.1.1
flask-cors==6.0.0
Flask-Login==0.6.3
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.165.0
google-auth==2.45.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.0
google-genai==1.56.0
google-generativeai==0.8.6
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
greenlet==3.2.2
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgpack==1.1.0
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
PyJWT==2.10.1
pymongo==4.15.5
pynvim==0.5.2
pyparsing==3.2.1
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
pytube @ git+https://github.com/pytube/pytube@a32fff39058a6f7e5e59ecd06a7467b71197ce35
requests==2.32.3
rsa==4.9
sniffio==1.3.1
soupsieve==2.7
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.1.1
urllib3==2.3.0
websockets==15.0.1
Werkzeug==3.1.3
